feat(gui): mount config and journal managers as built-in default kfx
Restructure the reference app into a shell plus two built-in default
extensions wired through a minimal internal kfx shape: a manifest that
declares the node-integrated (zero-copy) runtime tier and a view that
receives capability handles from the host instead of reaching for
globals. The config manager does location-scoped CRUD through the
runtime's config store; the journal manager browses recorded runs,
registered locations, and the merged event stream with a message-type
filter. The shape mirrors the intended loose kfx contract so
externalizing it later is a move, not a rewrite.
